Splic code by any '.rb' files
