+++ IvanVaskov/homework02/main/data.csv
@@ -1,27 +1,27 @@
 Cur_Abbreviation,Cur_Scale,Cur_Name,Cur_OfficialRate
-AUD,1,Австралийский доллар,1.7211
-BGN,1,Болгарский лев,1.4453
-UAH,100,Гривен,8.7014
-DKK,10,Датских крон,3.798
-USD,1,Доллар США,2.4103
-EUR,1,Евро,2.8281
-PLN,10,Злотых,6.421
-IRR,100000,Иранских риалов,5.7388
-ISK,100,Исландских крон,1.7885
-JPY,100,Йен,2.2895
-CAD,1,Канадский доллар,1.8013
-CNY,10,Китайских юаней,3.4418
-KWD,1,Кувейтский динар,7.8755
-MDL,10,Молдавских леев,1.4368
-NZD,1,Новозеландский доллар,1.602
-NOK,10,Норвежских крон,2.6413
-RUB,100,Российских рублей,3.3511
-XDR,1,СДР (Специальные права заимствования),3.3861
-SGD,1,Сингапурcкий доллар,1.7464
-KGS,100,Сомов,3.1446
-KZT,1000,Тенге,5.8054
-TRY,10,Турецких лир,3.4695
-GBP,1,Фунт стерлингов,3.1027
-CZK,100,Чешских крон,10.7778
-SEK,10,Шведских крон,2.7483
-CHF,1,Швейцарский франк,2.6222
+AUD,1,Австралийский доллар,1.7359
+BGN,1,Болгарский лев,1.4523
+UAH,100,Гривен,8.7371
+DKK,10,Датских крон,3.8166
+USD,1,Доллар США,2.418
+EUR,1,Евро,2.8413
+PLN,10,Злотых,6.4266
+IRR,100000,Иранских риалов,5.7571
+ISK,100,Исландских крон,1.791
+JPY,100,Йен,2.3042
+CAD,1,Канадский доллар,1.8116
+CNY,10,Китайских юаней,3.4557
+KWD,1,Кувейтский динар,7.9058
+MDL,10,Молдавских леев,1.4444
+NZD,1,Новозеландский доллар,1.6103
+NOK,10,Норвежских крон,2.6633
+RUB,100,Российских рублей,3.3462
+XDR,1,СДР (Специальные права заимствования),3.3986
+SGD,1,Сингапурcкий доллар,1.7587
+KGS,100,Сомов,3.1546
+KZT,1000,Тенге,5.787
+TRY,10,Турецких лир,3.473
+GBP,1,Фунт стерлингов,3.1358
+CZK,100,Чешских крон,10.8154
+SEK,10,Шведских крон,2.7624
+CHF,1,Швейцарский франк,2.6402
