+++ IvanVaskov/homework02/main/data.json
@@ -1,1 +1,1 @@
-{"AUD":{"Cur_Scale":1,"Cur_Name":"Австралийский доллар","Cur_OfficialRate":1.7211},"BGN":{"Cur_Scale":1,"Cur_Name":"Болгарский лев","Cur_OfficialRate":1.4453},"UAH":{"Cur_Scale":100,"Cur_Name":"Гривен","Cur_OfficialRate":8.7014},"DKK":{"Cur_Scale":10,"Cur_Name":"Датских крон","Cur_OfficialRate":3.798},"USD":{"Cur_Scale":1,"Cur_Name":"Доллар США","Cur_OfficialRate":2.4103},"EUR":{"Cur_Scale":1,"Cur_Name":"Евро","Cur_OfficialRate":2.8281},"PLN":{"Cur_Scale":10,"Cur_Name":"Злотых","Cur_OfficialRate":6.421},"IRR":{"Cur_Scale":100000,"Cur_Name":"Иранских риалов","Cur_OfficialRate":5.7388},"ISK":{"Cur_Scale":100,"Cur_Name":"Исландских крон","Cur_OfficialRate":1.7885},"JPY":{"Cur_Scale":100,"Cur_Name":"Йен","Cur_OfficialRate":2.2895},"CAD":{"Cur_Scale":1,"Cur_Name":"Канадский доллар","Cur_OfficialRate":1.8013},"CNY":{"Cur_Scale":10,"Cur_Name":"Китайских юаней","Cur_OfficialRate":3.4418},"KWD":{"Cur_Scale":1,"Cur_Name":"Кувейтский динар","Cur_OfficialRate":7.8755},"MDL":{"Cur_Scale":10,"Cur_Name":"Молдавских леев","Cur_OfficialRate":1.4368},"NZD":{"Cur_Scale":1,"Cur_Name":"Новозеландский доллар","Cur_OfficialRate":1.602},"NOK":{"Cur_Scale":10,"Cur_Name":"Норвежских крон","Cur_OfficialRate":2.6413},"RUB":{"Cur_Scale":100,"Cur_Name":"Российских рублей","Cur_OfficialRate":3.3511},"XDR":{"Cur_Scale":1,"Cur_Name":"СДР (Специальные права заимствования)","Cur_OfficialRate":3.3861},"SGD":{"Cur_Scale":1,"Cur_Name":"Сингапурcкий доллар","Cur_OfficialRate":1.7464},"KGS":{"Cur_Scale":100,"Cur_Name":"Сомов","Cur_OfficialRate":3.1446},"KZT":{"Cur_Scale":1000,"Cur_Name":"Тенге","Cur_OfficialRate":5.8054},"TRY":{"Cur_Scale":10,"Cur_Name":"Турецких лир","Cur_OfficialRate":3.4695},"GBP":{"Cur_Scale":1,"Cur_Name":"Фунт стерлингов","Cur_OfficialRate":3.1027},"CZK":{"Cur_Scale":100,"Cur_Name":"Чешских крон","Cur_OfficialRate":10.7778},"SEK":{"Cur_Scale":10,"Cur_Name":"Шведских крон","Cur_OfficialRate":2.7483},"CHF":{"Cur_Scale":1,"Cur_Name":"Швейцарский франк","Cur_OfficialRate":2.6222}}
+{"AUD":{"Cur_Scale":1,"Cur_Name":"Австралийский доллар","Cur_OfficialRate":1.7359},"BGN":{"Cur_Scale":1,"Cur_Name":"Болгарский лев","Cur_OfficialRate":1.4523},"UAH":{"Cur_Scale":100,"Cur_Name":"Гривен","Cur_OfficialRate":8.7371},"DKK":{"Cur_Scale":10,"Cur_Name":"Датских крон","Cur_OfficialRate":3.8166},"USD":{"Cur_Scale":1,"Cur_Name":"Доллар США","Cur_OfficialRate":2.418},"EUR":{"Cur_Scale":1,"Cur_Name":"Евро","Cur_OfficialRate":2.8413},"PLN":{"Cur_Scale":10,"Cur_Name":"Злотых","Cur_OfficialRate":6.4266},"IRR":{"Cur_Scale":100000,"Cur_Name":"Иранских риалов","Cur_OfficialRate":5.7571},"ISK":{"Cur_Scale":100,"Cur_Name":"Исландских крон","Cur_OfficialRate":1.791},"JPY":{"Cur_Scale":100,"Cur_Name":"Йен","Cur_OfficialRate":2.3042},"CAD":{"Cur_Scale":1,"Cur_Name":"Канадский доллар","Cur_OfficialRate":1.8116},"CNY":{"Cur_Scale":10,"Cur_Name":"Китайских юаней","Cur_OfficialRate":3.4557},"KWD":{"Cur_Scale":1,"Cur_Name":"Кувейтский динар","Cur_OfficialRate":7.9058},"MDL":{"Cur_Scale":10,"Cur_Name":"Молдавских леев","Cur_OfficialRate":1.4444},"NZD":{"Cur_Scale":1,"Cur_Name":"Новозеландский доллар","Cur_OfficialRate":1.6103},"NOK":{"Cur_Scale":10,"Cur_Name":"Норвежских крон","Cur_OfficialRate":2.6633},"RUB":{"Cur_Scale":100,"Cur_Name":"Российских рублей","Cur_OfficialRate":3.3462},"XDR":{"Cur_Scale":1,"Cur_Name":"СДР (Специальные права заимствования)","Cur_OfficialRate":3.3986},"SGD":{"Cur_Scale":1,"Cur_Name":"Сингапурcкий доллар","Cur_OfficialRate":1.7587},"KGS":{"Cur_Scale":100,"Cur_Name":"Сомов","Cur_OfficialRate":3.1546},"KZT":{"Cur_Scale":1000,"Cur_Name":"Тенге","Cur_OfficialRate":5.787},"TRY":{"Cur_Scale":10,"Cur_Name":"Турецких лир","Cur_OfficialRate":3.473},"GBP":{"Cur_Scale":1,"Cur_Name":"Фунт стерлингов","Cur_OfficialRate":3.1358},"CZK":{"Cur_Scale":100,"Cur_Name":"Чешских крон","Cur_OfficialRate":10.8154},"SEK":{"Cur_Scale":10,"Cur_Name":"Шведских крон","Cur_OfficialRate":2.7624},"CHF":{"Cur_Scale":1,"Cur_Name":"Швейцарский франк","Cur_OfficialRate":2.6402}}
